feat(crypto): consolidate dormant page envelopes into reddb-io-crypto (#1053)
Create the reddb-io-crypto authority crate (lib reddb_crypto) homing the
canonical per-page encryption-at-rest envelope, the mandatory encrypt
parameters, and key parsing — paralleling reddb-io-file / reddb-io-wire
under ADR 0046 (new ADR 0054).

Two dormant, byte-incompatible envelopes existed for the same not-yet-shipped
feature. The leaner magic-less frame survives as canonical (nonce(12) ‖ ct ‖
tag(16), 28-byte overhead, AAD = page_id u32 LE): it is already embedded in
the page-0 RDBE key_check blob (reddb_file's fixed 60-byte slot) and wired
into the dormant pager. The self-describing RDEP frame is retired — not
blind-deleted: its typed error enum, OS-CSPRNG nonce source, and hex/base64
key parser are carried forward into the new crate.

reddb-server keeps delegating shims (crypto::page_encryption facade with the
server-specific key_from_env; PageEncryptor binds a zeroizing SecureKey to the
canonical free functions). The on-disk frame is byte-identical to the prior
PageEncryptor output, so the dormant wiring and page-0 key_check are unchanged.

Validated: cargo check/build green (crypto + server + umbrella); reddb-io-crypto
13 tests (unit + format-pinning: overhead/field-order/AAD-endianness);
server-lib storage::encryption 19 + crypto::page_encryption 1; clippy/fmt clean
on touched code.
