+++ crates/reddb-crypto/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "reddb-io-crypto"
+version = "1.0.0"
+edition = "2021"
+authors = ["RedDB.io <hello@reddb.io>"]
+description = "RedDB cryptographic authority: the canonical per-page encryption-at-rest envelope (AES-256-GCM), mandatory encrypt parameters, and key parsing."
+license = "AGPL-3.0-only"
+homepage = "https://github.com/reddb-io/reddb"
+repository = "https://github.com/reddb-io/reddb"
+documentation = "https://docs.rs/reddb-io-crypto"
+keywords = ["database", "encryption", "aes-gcm", "at-rest", "envelope"]
+categories = ["cryptography", "database-implementations"]
+
+[lib]
+name = "reddb_crypto"
+path = "src/lib.rs"
+
+[dependencies]
+aes-gcm = "0.10"
