+++ crates/reddb-server/src/crypto/mod.rs
@@ -11,9 +11,8 @@ pub use aes_gcm::{aes256_gcm_decrypt, aes256_gcm_encrypt};
 pub use const_time::constant_time_eq;
 pub use hmac::hmac_sha256;
 pub use page_encryption::{
-    decrypt_page, encrypt_page, is_encrypted_frame, key_from_env,
-    parse_key as parse_encryption_key, PageEncryptionError, FRAME_MAGIC, FRAME_OVERHEAD,
-    FRAME_VERSION,
+    decrypt_page, encrypt_page, key_from_env, parse_key as parse_encryption_key, PageEnvelopeError,
+    PAGE_ENVELOPE_OVERHEAD,
 };
 pub use sha256::{sha256, Sha256};
 pub use uuid::Uuid;
